requirements.txt
astroid==2.3.2
autopep8==1.4.4
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
wrapt==1.11.2
